Refactor test fixtures and update consistency tests for HNetGRU and AttentionLayer
- Rename fixture model to hnetgru in conftest.py for clarity
- Update tests_consistency_hnet_gru.py to use the renamed hnetgru fixture
- Add test_AttentionLayer_init to tests_consistency_attention_layer.py
- Remove tests from tests_scenarios_attention_layer.py
- isort + black
