+++ tests/consistency_tests/run/tests_consistency_run.py
@@ -1,3 +1,1 @@
 # tests/consistency_tests/run/tests_consistency_run.py
-
-
